Project/wordfrequency/tests/tests_simplify_contextual_example/requirements.txt
pytest==7.3.1
requests==2.28.2
python-dotenv==1.0.0
textstat==0.7.3
Flask==2.2.2
typing==3.10.0.2
numpy==1.23.2
openai==0.27.4
